Change reset threshold to 3 seconds instead of 10 for bgm
Basically, sometimes the music never resets for some reason
I dont know what is happening. Update functions are weird.
So I make it check every frame for the song position vs the song length. If we pass that, we should restart the song. I dont know of a case where it gets too far really, so just let it be 3 seconds. This shouldnt cause issues.